+++ models/user.js
@@ -2,7 +2,7 @@ const mongoose = require('mongoose');
 const validator = require('validator');
 const bcrypt = require('bcryptjs');
 const uniqueValidator = require('mongoose-unique-validator');
-const { validOptions } = require('../libs/validOPtions');
+const validOptions = require('../libs/valid-options');
 const { getResponse } = require('../libs/helpers');
 const updateOptions = require('../libs/optionsForModeUpdatel');
 const RequestWrong = require('../errors/request-wrong');
